refactor(assert): Normalize interfaces signature and messages
